Basic project setup.
core is the frontend webapp.
sequencer is a microservice for finding mutations in a gene sequence.
drug_finder is a microservice for finding drugs that treat a given set of mutations.
drug_optimizer is a microservice for figuring out the best drug given a series of drugs that treat mutations.